requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
idna==3.2
postal==1.1.9
python-dotenv==0.19.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
webdriver-manager==3.4.2
